requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-vision-face==0.6.0
azure-common==1.1.28
boto3==1.23.10
botocore==1.26.10
cachelib==0.7.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
Flask-Login==0.6.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
msrest==0.6.21
oauthlib==3.2.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.36
urllib3==1.26.9
watchdog==2.1.8
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
